Fix underflow errors for tokens
Remove iterating over a list of tuples
test for now to verify the build is still
functioning.

- Show msg in gen
- Re-add logic for expected.
  Specifically, primitives disguised
  as expressions.
  This should be caught during the
  generation stage, but I'm not sure
  why it isn't.
  Need to investigate.

+++ src/common/position.rs
@@ -79,7 +79,7 @@ impl CaretPos {
     /// Create new [EndPoint] which is offset in the vertical direction by the
     /// given amount.
     pub fn offset_line(self, offset: usize) -> CaretPos {
-        CaretPos { line: self.line + offset, pos: self.pos }
+        CaretPos { line: (self.line as i32 + offset as i32) as usize, pos: self.pos }
     }
 
     /// Create new [EndPoint] which is offset in the horizontal direction by the
